stats-ai-analyzer/requirements.txt
flask==3.0.0
openai==1.12.0
PyPDF2==3.0.1
python-dotenv==1.0.0
Werkzeug==3.0.1
pdf2image==1.16.3
Pillow>=10.0.0
